+++ packages/stream_chat_flutter/lib/src/message_modal/message_actions_modal.dart
@@ -106,16 +106,18 @@ class StreamMessageActionsModal extends StatelessWidget {
         ),
     };
 
+    final alignment = switch (reverse) {
+      true => AlignmentDirectional.centerEnd,
+      false => AlignmentDirectional.centerStart,
+    };
+
     return StreamMessageModal(
-      alignment: switch (reverse) {
-        true => AlignmentDirectional.centerEnd,
-        false => AlignmentDirectional.centerStart,
-      },
+      alignment: alignment,
       headerBuilder: (context) {
         return Column(
           spacing: 10,
           mainAxisSize: MainAxisSize.min,
-          crossAxisAlignment: CrossAxisAlignment.stretch,
+          crossAxisAlignment: alignment.toColumnCrossAxisAlignment(context),
           children: <Widget?>[
             reactionPicker,
             IgnorePointer(child: messageWidget),
