+++ packages/stream_chat_flutter/lib/src/message_modal/message_reactions_modal.dart
@@ -102,16 +102,18 @@ class StreamMessageReactionsModal extends StatelessWidget {
         ),
     };
 
+    final alignment = switch (reverse) {
+      true => AlignmentDirectional.centerEnd,
+      false => AlignmentDirectional.centerStart,
+    };
+
     return StreamMessageModal(
-      alignment: switch (reverse) {
-        true => AlignmentDirectional.centerEnd,
-        false => AlignmentDirectional.centerStart,
-      },
+      alignment: alignment,
       headerBuilder: (context) {
         return Column(
           spacing: 10,
           mainAxisSize: MainAxisSize.min,
-          crossAxisAlignment: CrossAxisAlignment.stretch,
+          crossAxisAlignment: alignment.toColumnCrossAxisAlignment(context),
           children: <Widget?>[
             reactionPicker,
             IgnorePointer(child: messageWidget),
